% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscreenshow.R
\name{subscreenshow}
\alias{subscreenshow}
\title{(ii) Visualization}
\usage{
subscreenshow(
  scresults,
  variable_importance = NULL,
  host = NULL,
  port = NULL,
  NiceNumbers = c(1, 1.5, 2, 4, 5, 6, 8, 10)
)
}
\arguments{
\item{scresults}{SubScreenResult object with results from a subscreencalc call}

\item{variable_importance}{add description}

\item{host}{host name or IP address for Shiny display}

\item{port}{port number for Shiny display}

\item{NiceNumbers}{list of numbers used for a 'nice' scale}
}
\description{
Start the Shiny based interactive visualization tool to show the subgroup results
generated by subscreencalc.
See and explore all subgroup results at one glance. Pick and chose a specific
subgroup, the level of combinations or a certain factor with its combinations.
Switch easily between different endpoint/target variables.
}
\keyword{analysis}
\keyword{subgroup}
\keyword{visualization}
